fix: 권장 시간 합산 오버플로 방지를 위해 int에서 long으로 변경 (#45)

+++ src/main/java/com/cotato/blankit/domain/recommendation/controller/RecommendationController.java
@@ -40,7 +40,7 @@ public class RecommendationController {
     public ApiResponse<TodayRecommendationResponse> getTodayRecommendation(
             @AuthenticationPrincipal CustomUserDetails userDetails
     ) {
-        int totalRecommendedMinutes = recommendationService.calculateTodayRecommendedMinutes(userDetails.getUserId());
+        long totalRecommendedMinutes = recommendationService.calculateTodayRecommendedMinutes(userDetails.getUserId());
         return ApiResponse.success(new TodayRecommendationResponse(
                 LocalDate.now(clock),
                 totalRecommendedMinutes,
